cart and coupon module complete

+++ src/main/java/com/food/Dto/Request/CreateItemRequestDto.java
@@ -26,7 +26,7 @@ public class CreateItemRequestDto {
     private List<String> itemIngredients;
 
     @JsonProperty("item_price")
-    private BigDecimal itemPrice;
+    private Double itemPrice;
 
     @JsonProperty("initial_item_quantity")
     private Integer itemQuantity;
